requirements.txt
google-cloud-firestore>=2.7.3
requests>=2.28.1
beautifulsoup4>=4.11.1
firebase-admin>=6.0.1